Moved anonymizer types to a dedicated folder
