+++ Anonymizer/Type/IpAnonymizer.php
@@ -10,9 +10,10 @@
  * @link      https://www.superbrave.nl/
  */
 
-namespace SuperBrave\GdprBundle\Anonymizer;
+namespace SuperBrave\GdprBundle\Anonymizer\Type;
 
 use InvalidArgumentException;
+use SuperBrave\GdprBundle\Anonymizer\AnonymizerInterface;
 
 class IpAnonymizer implements AnonymizerInterface
 {
